Section 20. File upload and download

+++ controllers/admin.js
@@ -1,5 +1,7 @@
-const { validationResult } = require('express-validator/check')
+const mongoose = require('mongoose');
+const fileHelper = require('../util/file');
 
+const { validationResult } = require('express-validator/check');
 const errorHandlerObjectWrapper = require('../util/errorHandlerObjectWrapper');
 
 const Product = require('../models/product');
@@ -17,28 +19,51 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
   const title = req.body.title;
-  const imageUrl = req.body.imageUrl;
+  const image = req.file;
   const price = req.body.price;
   const description = req.body.description;
-  const { errors } = validationResult(req);
-
-  if (errors.length > 0) {
+  if (!image) {
     return res.status(422).render('admin/edit-product', {
       pageTitle: 'Add Product',
       path: '/admin/add-product',
       editing: false,
-      product: { title, imageUrl, price, description },
       hasError: true,
-      errorMessage: errors[0].msg,
-      validationErrors: errors,
+      product: {
+        title: title,
+        price: price,
+        description: description,
+      },
+      errorMessage: 'Attached file is not an image.',
+      validationErrors: [],
     });
   }
+  const errors = validationResult(req);
+
+  if (!errors.isEmpty()) {
+    console.log(errors.array());
+    return res.status(422).render('admin/edit-product', {
+      pageTitle: 'Add Product',
+      path: '/admin/add-product',
+      editing: false,
+      hasError: true,
+      product: {
+        title: title,
+        imageUrl: imageUrl,
+        price: price,
+        description: description,
+      },
+      errorMessage: errors.array()[0].msg,
+      validationErrors: errors.array(),
+    });
+  }
+
+  const imageUrl = image.path;
 
   const product = new Product({
     title: title,
     price: price,
     description: description,
-    imageUrl: imageUrl,
+    imageUrl,
     userId: req.user
   });
   product
@@ -53,13 +78,20 @@ exports.postAddProduct = (req, res, next) => {
       //   pageTitle: 'Add Product',
       //   path: '/admin/add-product',
       //   editing: false,
-      //   product: { title, imageUrl, price, description },
       //   hasError: true,
-      //   errorMessage: 'Database operation failed.',
-      //   validationErrors: [],
+      //   product: {
+      //     title: title,
+      //     imageUrl: imageUrl,
+      //     price: price,
+      //     description: description
+      //   },
+      //   errorMessage: 'Database operation failed, please try again.',
+      //   validationErrors: []
       // });
       // res.redirect('/500');
-      return errorHandlerObjectWrapper(500, err);
+      const error = new Error(err);
+      error.httpStatusCode = 500;
+      return next(error);
     });
 };
 
@@ -91,25 +123,25 @@ exports.postEditProduct = (req, res, next) => {
   const prodId = req.body.productId;
   const updatedTitle = req.body.title;
   const updatedPrice = req.body.price;
-  const updatedImageUrl = req.body.imageUrl;
+  const image = req.file;
   const updatedDesc = req.body.description;
-  const { errors } = validationResult(req);
 
-  if (errors.length > 0) {
+  const errors = validationResult(req);
+
+  if (!errors.isEmpty()) {
     return res.status(422).render('admin/edit-product', {
       pageTitle: 'Edit Product',
       path: '/admin/edit-product',
       editing: true,
+      hasError: true,
       product: {
         title: updatedTitle,
-        imageUrl: updatedImageUrl,
         price: updatedPrice,
         description: updatedDesc,
         _id: prodId,
       },
-      hasError: true,
-      errorMessage: errors[0].msg,
-      validationErrors: errors,
+      errorMessage: errors.array()[0].msg,
+      validationErrors: errors.array(),
     });
   }
 
@@ -121,12 +153,14 @@ exports.postEditProduct = (req, res, next) => {
       product.title = updatedTitle;
       product.price = updatedPrice;
       product.description = updatedDesc;
-      product.imageUrl = updatedImageUrl;
-      return product.save()
-        .then(result => {
-          console.log('UPDATED PRODUCT!');
-          res.redirect('/admin/products');
-        });
+      if (image) {
+        fileHelper.deleteFile(product.imageUrl);
+        product.imageUrl = image.path;
+      }
+      return product.save().then(result => {
+        console.log('UPDATED PRODUCT!');
+        res.redirect('/admin/products');
+      });
     })
     .catch(err => errorHandlerObjectWrapper(500, err));
 };
@@ -148,13 +182,16 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({
-    _id: prodId,
-    userId: req.user._id,
+  Product.findById(prodId).then(product => {
+    if (!product) {
+      next(new Error('Product not found!'));
+    }
+    fileHelper.deleteFile(product.imageUrl);
+    Product.deleteOne({ _id: prodId, userId: req.user._id });
   })
-    .then(() => {
-      console.log('DESTROYED PRODUCT');
-      res.redirect('/admin/products');
-    })
-    .catch(err => errorHandlerObjectWrapper(500, err));
+  .then(() => {
+    console.log('DESTROYED PRODUCT');
+    res.redirect('/admin/products');
+  })
+  .catch(err => errorHandlerObjectWrapper(500, err));
 };
